+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/DependencyGraphChecker.kt
@@ -30,6 +30,8 @@ import dev.zacsweers.metro.compiler.fir.validateBindingRef
 import dev.zacsweers.metro.compiler.fir.validateInjectionSiteType
 import dev.zacsweers.metro.compiler.mapToSet
 import dev.zacsweers.metro.compiler.metroAnnotations
+import dev.zacsweers.metro.compiler.tracing.TraceCategories
+import dev.zacsweers.metro.compiler.tracing.trace
 import org.jetbrains.kotlin.KtSourceElement
 import org.jetbrains.kotlin.descriptors.Modality
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
@@ -70,7 +72,12 @@ internal object DependencyGraphChecker : FirClassChecker(MppCheckerKind.Common) 
 
   context(context: CheckerContext, reporter: DiagnosticReporter)
   override fun check(declaration: FirClass) {
-    context(context.session.compatContext) { checkImpl(declaration) }
+    context.session.trace(
+      name = "DependencyGraphChecker(${declaration.classId})",
+      category = TraceCategories.FIR_CHECKER,
+    ) {
+      context(context.session.compatContext) { checkImpl(declaration) }
+    }
   }
 
   context(context: CheckerContext, reporter: DiagnosticReporter, compatContext: CompatContext)