+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/tracing/EmptyTraceScope.kt
@@ -1,15 +1,14 @@
 // Copyright (C) 2026 Zac Sweers
 // SPDX-License-Identifier: Apache-2.0
-package dev.zacsweers.metro.compiler
+package dev.zacsweers.metro.compiler.tracing
 
 import androidx.tracing.AbstractTraceSink
 import androidx.tracing.DelicateTracingApi
 import androidx.tracing.PooledTracePacketArray
 import androidx.tracing.wire.TraceDriver
-import dev.zacsweers.metro.compiler.tracing.TraceScope
 
-internal fun testTraceScope(): TraceScope {
-  return TraceScope(TraceDriver(EmptyTraceSink, isEnabled = false).tracer, "test")
+internal fun emptyTraceScope(category: String): TraceScope {
+  return TraceScope(TraceDriver(EmptyTraceSink, isEnabled = false).tracer, category)
 }
 
 private object EmptyTraceSink : AbstractTraceSink() {