+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/graph/MetroSortTest.kt
@@ -20,8 +20,8 @@ package dev.zacsweers.metro.compiler.graph
 
 import androidx.collection.MutableScatterMap
 import com.google.common.truth.Truth.assertThat
-import dev.zacsweers.metro.compiler.testTraceScope
 import dev.zacsweers.metro.compiler.tracing.TraceScope
+import dev.zacsweers.metro.compiler.tracing.emptyTraceScope
 import java.util.SortedSet
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -31,7 +31,7 @@ import kotlin.test.assertTrue
 import kotlin.test.fail
 import org.jetbrains.annotations.TestOnly
 
-class MetroSortTest : TraceScope by testTraceScope() {
+class MetroSortTest : TraceScope by emptyTraceScope("\"test\"") {
   @Test
   fun emptyEdges() {
     val unsorted = listOf("a", "b", "c")
@@ -664,7 +664,7 @@ private fun <T : Comparable<T>> Iterable<T>.topologicalSort(
       }
     }
   val fullAdjacency = buildFullAdjacency(fakeMap, sourceToTarget, onMissing)
-  val topology = with(testTraceScope()) { metroSort(fullAdjacency, isDeferrable, onCycle) }
+  val topology = with(emptyTraceScope("test")) { metroSort(fullAdjacency, isDeferrable, onCycle) }
   return topology.sortedKeys
 }